fix score_id

+++ src/database/migrations/2020_04_27_080508_add_columns_to_users_table.php
@@ -25,7 +25,7 @@ class AddColumnsToUsersTable extends Migration
             $table->boolean('become_zerone_member')->after('job_title')->nullable();
             $table->boolean('zerone_subscriptions')->after('become_zerone_member')->nullable();
             $table->boolean('zerone_breaking_news')->after('zerone_subscriptions')->nullable();
-            $table->unsignedTinyInteger('score_id')->after('zerone_breaking_news')->nullable();
+            $table->unsignedBigInteger('score_id')->after('zerone_breaking_news')->nullable();
             $table->unsignedTinyInteger('is_black_list')->after('score_id')->nullable();
         });
     }
